Add array/tree-traversing branch to waitForElementByClass().
This modifies waitForElementByClass() to take a parent in which to search, and allow elementClass can be an array.  If it's an array, it will work through the array looking for successive children starting from the original parent.  Thus we can wait for .title-qAcLxz, and then look for the .flex-lFgbSz within it.